+++ common/platform/windows/SysFile.hpp
@@ -129,7 +129,7 @@ public:
 
 protected:
     void   getStreamTypeInfo();
-    int64_t writeData(const char *data, size_t length);
+    size_t writeData(const char *data, size_t length);
 
     int    fileHandle;      // separate file handle
     int    errInfo;         // last error info
